+++ zcash_client_backend/src/fees/zip317.rs
@@ -16,8 +16,8 @@ use zcash_primitives::{
 use crate::ShieldedProtocol;
 
 use super::{
-    common::single_change_output_balance, sapling as sapling_fees, ChangeError, ChangeStrategy,
-    DustOutputPolicy, TransactionBalance,
+    common::{calculate_net_flows, single_change_output_balance, single_change_output_policy},
+    sapling as sapling_fees, ChangeError, ChangeStrategy, DustOutputPolicy, TransactionBalance,
 };
 
 #[cfg(feature = "orchard")]
@@ -93,32 +93,73 @@ impl ChangeStrategy for SingleOutputChangeStrategy {
             })
             .collect();
 
+        #[cfg(feature = "orchard")]
+        let mut orchard_dust: Vec<NoteRefT> = orchard
+            .inputs()
+            .iter()
+            .filter_map(|i| {
+                if orchard_fees::InputView::<NoteRefT>::value(i) < self.fee_rule.marginal_fee() {
+                    Some(orchard_fees::InputView::<NoteRefT>::note_id(i).clone())
+                } else {
+                    None
+                }
+            })
+            .collect();
+        #[cfg(not(feature = "orchard"))]
+        let mut orchard_dust: Vec<NoteRefT> = vec![];
+
         // Depending on the shape of the transaction, we may be able to spend up to
         // `grace_actions - 1` dust inputs. If we don't have any dust inputs though,
         // we don't need to worry about any of that.
-        if !(transparent_dust.is_empty() && sapling_dust.is_empty()) {
+        if !(transparent_dust.is_empty() && sapling_dust.is_empty() && orchard_dust.is_empty()) {
             let t_non_dust = transparent_inputs.len() - transparent_dust.len();
             let t_allowed_dust = transparent_outputs.len().saturating_sub(t_non_dust);
 
-            // We add one to the sapling outputs for the (single) change output Note that this
-            // means that wallet-internal shielding transactions are an opportunity to spend a dust
-            // note.
+            // We add one to either the Sapling or Orchard outputs for the (single)
+            // change output. Note that this means that wallet-internal shielding
+            // transactions are an opportunity to spend a dust note.
+            let net_flows = calculate_net_flows::<NoteRefT, Self::FeeRule, Self::Error>(
+                transparent_inputs,
+                transparent_outputs,
+                sapling,
+                #[cfg(feature = "orchard")]
+                orchard,
+            )?;
+            let (_, sapling_change, orchard_change) =
+                single_change_output_policy::<NoteRefT, Self::FeeRule, Self::Error>(
+                    &net_flows,
+                    self.fallback_change_pool,
+                )?;
+
             let s_non_dust = sapling.inputs().len() - sapling_dust.len();
-            let s_allowed_dust = (sapling.outputs().len() + 1).saturating_sub(s_non_dust);
+            let s_allowed_dust =
+                (sapling.outputs().len() + sapling_change).saturating_sub(s_non_dust);
+
+            #[cfg(feature = "orchard")]
+            let (orchard_inputs_len, orchard_outputs_len) =
+                (orchard.inputs().len(), orchard.outputs().len());
+            #[cfg(not(feature = "orchard"))]
+            let (orchard_inputs_len, orchard_outputs_len) = (0, 0);
+
+            let o_non_dust = orchard_inputs_len - orchard_dust.len();
+            let o_allowed_dust = (orchard_outputs_len + orchard_change).saturating_sub(o_non_dust);
 
             let available_grace_inputs = self
                 .fee_rule
                 .grace_actions()
                 .saturating_sub(t_non_dust)
-                .saturating_sub(s_non_dust);
+                .saturating_sub(s_non_dust)
+                .saturating_sub(o_non_dust);
 
             let mut t_disallowed_dust = transparent_dust.len().saturating_sub(t_allowed_dust);
             let mut s_disallowed_dust = sapling_dust.len().saturating_sub(s_allowed_dust);
+            let mut o_disallowed_dust = orchard_dust.len().saturating_sub(o_allowed_dust);
 
             if available_grace_inputs > 0 {
                 // If we have available grace inputs, allocate them first to transparent dust
-                // and then to Sapling dust. The caller has provided inputs that it is willing
-                // to spend, so we don't need to consider privacy effects at this layer.
+                // and then to Sapling dust followed by Orchard dust. The caller has provided
+                // inputs that it is willing to spend, so we don't need to consider privacy
+                // effects at this layer.
                 let t_grace_dust = available_grace_inputs.saturating_sub(t_disallowed_dust);
                 t_disallowed_dust = t_disallowed_dust.saturating_sub(t_grace_dust);
 
@@ -126,6 +167,12 @@ impl ChangeStrategy for SingleOutputChangeStrategy {
                     .saturating_sub(t_grace_dust)
                     .saturating_sub(s_disallowed_dust);
                 s_disallowed_dust = s_disallowed_dust.saturating_sub(s_grace_dust);
+
+                let o_grace_dust = available_grace_inputs
+                    .saturating_sub(t_grace_dust)
+                    .saturating_sub(s_grace_dust)
+                    .saturating_sub(o_disallowed_dust);
+                o_disallowed_dust = o_disallowed_dust.saturating_sub(o_grace_dust);
             }
 
             // Truncate the lists of inputs to be disregarded in input selection to just the
@@ -135,11 +182,16 @@ impl ChangeStrategy for SingleOutputChangeStrategy {
             transparent_dust.truncate(t_disallowed_dust);
             sapling_dust.reverse();
             sapling_dust.truncate(s_disallowed_dust);
+            orchard_dust.reverse();
+            orchard_dust.truncate(o_disallowed_dust);
 
-            if !(transparent_dust.is_empty() && sapling_dust.is_empty()) {
+            if !(transparent_dust.is_empty() && sapling_dust.is_empty() && orchard_dust.is_empty())
+            {
                 return Err(ChangeError::DustInputs {
                     transparent: transparent_dust,
                     sapling: sapling_dust,
+                    #[cfg(feature = "orchard")]
+                    orchard: orchard_dust,
                 });
             }
         }